requirements.txt
# Core dependencies for OrderOfBooks scraper and matcher
playwright>=1.40.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
python-slugify>=8.0.0
rapidfuzz>=3.5.0
openai>=1.3.0

# Spanish edition finder dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
